Makefile: use traditional DESTDIR/PREFIX conventions
DESTDIR is normally used by distribution packagers to install to a
path that the final application will not be located in.  Particularly
handy for building packages as a limited user.

Instead, use PREFIX to actually determine the install location on the
system, and default to /usr/local.  Manually installing software to
/usr is typically bad practice, stomping over system package
management, or system package management stomping over you.  PREFIX
should usually only be set to /usr when part of a package build
script.